Renamed serialization/normalization services
