+++ apps/api/src/Application/Normalization/Normalizer/MessageNormalizer.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace Application\Normalizer\Internal;
+declare(strict_types=1);
 
-use Application\Normalizer\Normalizer;
+namespace Application\Normalization\Normalizer;
+
+use Application\Normalization\Normalizer;
 use Domain\Model\Message;
 
+/**
+ * @implements Normalizer<Message>
+ */
 final class MessageNormalizer implements Normalizer
 {
-    /**
-     * {@inheritDoc}
-     */
     public function supports(mixed $data): bool
     {
         return $data instanceof Message;
@@ -22,7 +24,7 @@ final class MessageNormalizer implements Normalizer
      *
      * @return array<string,string|array<string,string>|null>
      */
-    public function normalize(mixed $data): array
+    public function normalize(mixed $data): mixed
     {
         return [
             'id' => (string) $data->getIdentifier(),
