+++ apps/api/src/Application/Normalization/Normalizer/StationNormalizer.php
@@ -1,11 +1,16 @@
 <?php
 
-namespace Application\Normalizer\Internal;
+declare(strict_types=1);
 
-use Application\Normalizer\Normalizer;
+namespace Application\Normalization\Normalizer;
+
+use Application\Normalization\Normalizer;
 use Domain\Model\Channel;
 use Domain\Model\Station;
 
+/**
+ * @implements Normalizer<Station>
+ */
 final class StationNormalizer implements Normalizer
 {
     /**
@@ -23,7 +28,7 @@ final class StationNormalizer implements Normalizer
      *
      * @return array<string,array<array<string,string>>|string|null>
      */
-    public function normalize(mixed $data): array
+    public function normalize(mixed $data): mixed
     {
         return [
             'id'   => (string) $data->getIdentifier(),
@@ -32,7 +37,7 @@ final class StationNormalizer implements Normalizer
             'createdAt' => $data->getCreatedAt()->format('Y-m-d'),
             'updatedAt' => $data->getUpdatedAt()->format('Y-m-d'),
             'channels' => array_map(
-                fn (Channel $channel): array => [
+                static fn (Channel $channel): array => [
                     'id' => (string) $channel->getIdentifier(),
                     'name' => $channel->getName()
                 ],
