--- apps/api/src/Application/Normalizer/Normalizer.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Application\Normalizer;
-
-use Application\Normalizer\Exception\NormalizerNotFoundException;
-
-interface Normalizer
-{
-    public function supports(mixed $data): bool;
-
-    /**
-     * @return mixed
-     *
-     * @throws NormalizerNotFoundException
-     */
-    public function normalize(mixed $data): mixed;
-}
